+++ src/models/db/bid.rs
@@ -1,4 +1,3 @@
-use bigdecimal::BigDecimal;
 use chrono::NaiveDateTime;
 use serde::{Deserialize, Serialize};
 
@@ -15,7 +14,7 @@ pub struct DbBid {
     pub market_name: Option<String>,
     pub nonce: Option<String>,
     pub nft_id: Option<String>,
-    pub price: Option<BigDecimal>,
+    pub price: Option<i64>,
     pub price_str: Option<String>,
     pub receiver: Option<String>,
     pub remaining_count: Option<i64>,
